% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImpTC.R
\name{ImpTC}
\alias{ImpTC}
\title{Non-circular Extended Nearest Insertion}
\usage{
ImpTC(Data, Seq, condn, Ndg=3)
}
\arguments{
\item{Data}{gene-by-sample matrix or isoform-by-sample matrix.It should be rescaled to gene specific z scores.}

\item{Seq}{NULL or a vector indicates the sample order.
if specified, the samples will be first reordered by this vector.}

\item{condn}{a numerical vector indicates conditions}

\item{Ndg}{degree of polynomial.}
}
\value{
This function performs the extended nearest insertion (ENI).
The ENI algorithm searchs for the optimal sample order which minimizes the MSE of
the polynomial regression  (PR).
This function will call PipeRCDF() function, which fits
PR to each row of the data.
The aggregated MSE of a fit is defined as the
summation of the MSEs of all genes/isoforms considered here.
The output returns the optimal order which provides the smallest PR MSE.
Note that the reordering happens within condition (time point). Cells from
different conditions won't be mixed unless the cell is placed in the
coundary of two conditions.
}
\description{
Non-circular Extended Nearest Insertion
}
\examples{
aa <- sin(seq(0,1,.1))
bb <- sin(seq(0.5,1.5,.1))
cc <- sin(seq(0.9,1.9,.1))
res <- ImpTC(rbind(aa,bb,cc), Seq=NULL, condn = c(rep(1,5),rep(2,6)))
}
\author{
Ning Leng
}

