% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Opt2TC.R
\name{Opt2TC}
\alias{Opt2TC}
\title{Opt-2 algorithm for non-circular Extended Nearest Insertion}
\usage{
Opt2TC(Data,N,Seq,condn, Ndg=3, NCThre=1000)
}
\arguments{
\item{Data}{gene-by-sample matrix or isoform-by-sample matrix.It should be r
escaled to gene specific z scores.}

\item{N,NCThre}{The 2-opt algorithm will stop if N iterations has been performed or if the optimal order
remains unchanged for over NCThre iterations.}

\item{Seq}{a vector indicates the sample order obtained from the ENI.}

\item{condn}{a numerical vector indicates conditions}

\item{Ndg}{degree of polynomial.}
}
\value{
This function performs the
the 2-opt algorithm to improve the optimal order searching of the Extended Nearest Insertion (ENI).
In each iteration, the function will randomly choose two points (samples), then flip the samples
between these two points. The new order will be adapted if it provides smaller PR MSE.
The output returns the optimal order and its PR MSE.
Note that the reordering happens within condition (time point). Cells from
different conditions won't be mixed unless the cell is placed in the
coundary of two conditions.
}
\description{
Opt-2 algorithm for non-circular Extended Nearest Insertion
}
\examples{
aa <- sin(seq(0,1,.1))
bb <- sin(seq(0.5,1.5,.1))
cc <- sin(seq(0.9,1.9,.1))
res <- ImpTC(rbind(aa,bb,cc), Seq=NULL, condn = c(rep(1,5),rep(2,6)))
res2 <- Opt2TC(rbind(aa,bb,cc), N=50, Seq=res, condn= c(rep(1,5),rep(2,6)))
}
\author{
Ning Leng
}

