% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rescale.R
\name{Rescale}
\alias{Rescale}
\title{Rescale the gene/isoform expression matrix}
\usage{
Rescale(Data)
}
\arguments{
\item{Data}{input gene-by-sample matrix or isoform-by-sample matrix}
}
\description{
Rescale the gene/isoform expression matrix
}
\note{
The output will be a gene-by-sample or isoform-by-sample matrix.
For each gene/isoform, the expressions will be scaled to mean 0 var 1
}
\examples{
Rescale(matrix(rnorm(10), nrow=2))
}
\author{
Ning Leng
}

