\name{WaveCrest-package}
\alias{WaveCrest-package}
\alias{WaveCrest}
\docType{package}
\title{
WaveCrest: a statistical approach to 
reconstruct gene expression trajectory in ordered single cell
RNA-seq experiments
}
\description{
WaveCrest is a statistical method that 
reconstructs gene expression trajectory in single cell RNA-seq experiments with ordered conditions.
WaveCrest contains two modules - the first module implements an extended nearest insertion algorithm that
searches for optimal cell orders, and the second module implements a spline fitting module
that could be used to identify additional dynamic genes.
}
\details{
	\tabular{ll}{
		Package: \tab WaveCrest\cr
		Type: \tab Package\cr
		License: \tab Artistic-2.0\cr
	}
}
\author{
	Ning Leng

	Maintainer: Ning Leng <lengning1@gmail.com>
}
\references{
}
\keyword{ package }
