% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WaveCrestIden.R
\name{WaveCrestIden}
\alias{WaveCrestIden}
\title{Identify additional dynamic genes based on recovered order}
\usage{
WaveCrestIden(Data, Order, Ndg=5, MeanLOD=10)
N=20000, NCThre=1000)
}
\arguments{
\item{Data}{gene-by-sample matrix or isoform-by-sample matrix. It should be rescaled to values bwteen
[-1,1].}

\item{Order}{A numeric vector indicates the reconstructed cell order.}

\item{Ndg}{degree of polynomial.}

\item{MeanLOD}{genes whose mean is less than MeanLOD are not considered.}
}
\value{
This function may be used to identify additional dynamic genes based on the recovered order.
}
\description{
Identify additional dynamic genes based on recovered order
}
\examples{
aa <- sin(seq(0,1,.1))
bb <- sin(seq(0.5,1.5,.1))
cc <- sin(seq(0.9,1.9,.1))
dd <- sin(seq(1.2,2.2,.1))
res <- WaveCrestENI(c("aa","bb"), rbind(aa,bb,cc,dd), N=50,
Conditions = factor(c(rep(1,5),rep(2,6))))
res.more <- WaveCrestIden(rbind(cc,dd), res, MeanLOD=0)
}
\author{
Ning Leng
}

